Handle missing values in CLI
Not all programmes have all information. Found an example where thte
`series_name` was absent.

Failure spotted with this URL:

https://www.bbc.co.uk/schedules/p00fzl8v/2021/01/30

It's plausible that the `name` or `description` might be absent, so
perform a similar check for those.

It shouldn't be possible for the start date or URL to be missing;
something is seriously broken with the page in that case.